backend/requirements.txt
alchemical==0.7.1
alembic==1.11.1
apifairy==1.3.0
apispec==6.3.0
blinker==1.6.2
click==8.1.3
Faker==18.9.0
Flask==2.3.2
Flask-Cors==3.0.10
Flask-HTTPAuth==4.8.0
flask-marshmallow==0.15.0
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-sqlalchemy==0.29.0
packaging==23.1
python-dateutil==2.8.2
python-dotenv==1.0.0
six==1.16.0
SQLAlchemy==2.0.14
typing_extensions==4.5.0
webargs==8.2.0
Werkzeug==2.3.4
